backend/requirements.txt
async-timeout==4.0.2
click==8.1.3
Deprecated==1.2.13
Flask==2.2.2
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pyparsing==3.0.9
python-dotenv==0.21.0
redis==4.3.4
rq==1.11.1
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
